Hexcode component refactoring
Hex component in its own file now and text is no longer a state value (instead, we keep track of whether the hex has been copied and modify the text displayed accordingly).

Related to pull request #29
